Rapid gun new weapon subclass for making weapons rapid fire
