+++ Assets/Scripts/WeaponScripts/Weapon.cs
@@ -1,6 +1,5 @@
 using UnityEngine;
 using MoreMountains.Feedbacks;
-using UnityEngine.InputSystem;
 
 public class Weapon : MonoBehaviour {
 
@@ -17,11 +16,8 @@ public class Weapon : MonoBehaviour {
 	[Header("Firing Vars")]
 	public int weaponLevel = 1;
 	public float rateOfFire = 2.0f;
-	private float originalRof;
-	public bool isRapidFiring;
-	public bool canFire = true;
+	[SerializeField] private bool canFire = true;
 	public float nextFire;
-	public float weaponCoolDown;
 	
 	[Header("Projectile Vars")]
 	public GameObject[] firePointArray;
@@ -34,20 +30,10 @@ public class Weapon : MonoBehaviour {
 
 	private void Start() {
 		audioSource = GetComponent<AudioSource>();
-		originalRof = rateOfFire;
 	}
 
 	private void FixedUpdate() {
 		CanFireTimer();
-
-		if (Time.time > weaponCoolDown) {
-			RapidFireTimer();
-			weaponCoolDown = Time.time + 2;
-		}
-			
-			
-		
-		
 	}
 	public virtual void FireWeapon(Vector3 firePoint, Vector3 targetPosition) {
 		audioSource.Play();
@@ -74,22 +60,5 @@ public class Weapon : MonoBehaviour {
 
 	public void AquireTarget(BaseEnemy enemy) {
 		enemyTarget = enemy.transform.position;
-	}
-
-	public void RapidFireTimer() {
-		bool RapidFireTimerTriggered = false;
-		
-			if (!isRapidFiring) {
-				rateOfFire = .15f;
-				isRapidFiring = true;
-				Debug.Log($"time: {Time.time}");
-				RapidFireTimerTriggered = true;
-			}
-		
-			if (isRapidFiring && !RapidFireTimerTriggered) {
-				rateOfFire = 1f;
-				isRapidFiring = false;
-				Debug.Log($"weaponCD: {weaponCoolDown}");
-			}
 	}
 }
